Generate and serve devcontainer proxy with a working `localhost` HTTPS SSL certification for HTTPS development (#35)
Pretty awesome change that I learned a lot from. This helps get
zoneblitz running in a dev container WITH HTTPS support.

This will allow us to eventually support any third party APIs or
services that require an HTTPS connection. It also helps ensure that
we're more consistent with production to help with any SSL issues on the
proxy in a dev environment.