Setup health check endpoint and monitor health in Docker containers (#49)
Sets up a `/health` GET endpoint on the server that the `Dockerfile`
uses to check health of the container. This will be useful for not only
the state of our deployments, but will make the deployment check less
flakey as we don't need to try HTTP requests over and over and know when
the container is healthy.